anvil_zksync_common

Module address_map

Structs

KnownAddress

Enums

ContractType

Statics

KNOWN_ADDRESSES

Loads the known contact addresses from the JSON file.

Functions

is_precompile

Checks if the given address is a precompile based on KNOWN_ADDRESSES.

is_system

Checks if the given address is a system contract based on KNOWN_ADDRESSES.